# API/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "P05Shop.API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UserLoginDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/StringServiceResponse"
            }
          }
        }
      }
    },
    "/api/Auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UserRegisterDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Int32ServiceResponse"
            }
          }
        }
      }
    },
    "/api/Auth/change-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/BooleanServiceResponse"
            }
          }
        }
      }
    },
...
    "/api/Book/createBook": {
      "post": {
        "tags": [
          "Book"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Book"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/BookListServiceResponse"
            }
          }
        }
      }
    },
    "/api/Book/DeleteBook/{id}": {
      "delete": {
        "tags": [
          "Book"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/BookListServiceResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Book": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "author": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "BookListServiceResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Book"
          }
        },
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "codeError": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "BooleanServiceResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "boolean"
        },
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "codeError": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
...
    "Int32ServiceResponse": {
      "type": "object",
      "properties": {
        "data": {
          "format": "int32",
          "type": "integer"
        },
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "codeError": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
...
    "StringServiceResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "string"
        },
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        },
        "codeError": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "UserLoginDTO": {
      "required": [
        "email",
        "password"
      ],
      "type": "object",
      "properties": {
        "email": {
          "minLength": 1,
          "type": "string"
        },
        "password": {
          "minLength": 1,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "UserRegisterDTO": {
      "required": [
        "email",
        "password",
        "username"
      ],
      "type": "object",
      "properties": {
        "email": {
          "format": "email",
          "minLength": 1,
          "type": "string"
        },
        "password": {
          "minLength": 6,
          "type": "string"
        },
        "confirmPassword": {
          "type": "string"
        },
        "username": {
          "minLength": 3,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
  }
}